Assign ports sequentially based on their formation order
Given the Procfile:

    alpha: ./alpha
    bravo: ./bravo

And the formation "bravo=1", the original implementation would assign
PORT => 8100 to process bravo.1. That was non-intuitive, so now ports
are assigned based on their position in the formation, completely
ignoring their Procfile order. In other words, a formation of
"bravo=1,alpha=2" with port 8000 would see the following assignments:

    bravo.1 => 8000
    alpha.1 => 8001
    alpha.2 => 8002